Improved version of rollingball.py
It now plots the different energies and it has a new definition for the angular velocity. It can be seen that total energy is conserved enough during most of the movement, but when zoomed in, the total energy shows a weird jump around t=0.7. I found out that this jump coincides with the ball being on the x position where the third derivative of the Gaussian function has a steep peak. So I think this might have something to do with it. 
If the number of time steps is increased, though, the difference gets significantly smaller (despite having the same shape).