Added new potentials. Updated Doc. Added animation.
New potentials are Woods-Saxon and Gaussian. Animation is disabled (the lines are commented as it requires ffmpeg to work). The definition of the potentials are now in the new file "potentials.py" to remove clutter from "particle.py" so both files are required. The "diagnostic" plots were also updated (now with x and y phase-space).