Reworked Woods-Saxon to make recangles and added it to UI
The parameters are:
- x0 and y0: position. These two parameters are now common to the gaussian potential and the WS potential.
- V0: potential depth/height
- Rx and Ry: 'radius' for x and y (half the lenght of the rectangle)
- a: slope, the smaller the more steep the potential will be.
I'm still fine tunning the allowed parameters for the WS potential (especially the parameter a). If you find that after computation and pressing 'Play' no animation appears it may be because the computation failed due to the parameters. If that happens a bunch of nan will appear on the console where you are executing the program, please let me know which values were used so I can tune the ranges.