More elctrons than targets + other modfications
Now you measure a number of electrons equal to the level + 1 except in level one that you only measure one electron. 
Bug in "play again" button found and solved, there was an issue with the lives in the new game. 
Now there are more probabilities of recovering a life in higher levels.
Sliders movement more fluid now, still room to improve.