Change in number of e- per level and optimization of computing time.
A part from level 1 (only one electron) in next levels we have 2n-1 electrons where n is the current level in which we are playing. 
Now when we plot the function the code runs faster (still not enough to make it plot the function every time we move one slider). It runs faster because now the eigenvalues and eigenvectors are found using that the matrix is tridiagonal, also, instead of using 500 points we only use 250. Additionally, when we chose the eigenvalue the program only computes it's eigenvalue and correspondent eigenvector instead of compuitng all eigenvalues and eigenvectors.