Bug detected and solved in executable
Bug detected in executable: some text labels weren't loaded correctly in the exe file and at first appeared with a different color, kind of greyish, when the label was called in the code changing it's size or text then it turned to black (the color the label should be from the beginning). It's strange because when the code is called from the py file without having an exe file this doesn't happen and all works correctly. 
The bug has been solved by changing the color from black to black(something completely unnecessary) so the label is called in the code.